requirements.txt
PyKakao==0.0.7
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.1
openpyxl==3.1.2
folium==0.16.0
tqdm==4.66.2
gensim==4.3.2
torch==2.2.2
scipy==1.8.0
openai==1.23.6
haversine=2.8.1